feat: using mixin and generic views

+++ store/views.py
@@ -3,63 +3,49 @@ from django.http import HttpResponse
 from django.db.models.aggregates import Count
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView
 from rest_framework import status
 from .models import Product, Collection
 from .serializers import ProductSerializer, CollectionSerializer
 
 # Create your views here.
 
-@api_view(['GET', 'POST'])
-def product_list(request):
-    if request.method == 'GET':
-        queryset = Product.objects.select_related('collection').all()[:5]
-        serializer = ProductSerializer(queryset, many=True, context={'request': request})
-        return Response(serializer.data)
-    elif request.method == 'POST':
-        serializer = ProductSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        print(request.data)
-        print(serializer.validated_data)
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-        
+class ProductList(ListCreateAPIView):
+    queryset = Product.objects.select_related('collection').all()[:5]
+    serializer_class = ProductSerializer
 
-@api_view(['GET', 'PUT', 'DELETE'])
-def product_details(request, id):
-    product = get_object_or_404(Product, pk=id)
-    if request.method == 'GET':
-        serializer = ProductSerializer(product)
-        print(serializer.data)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = ProductSerializer(product, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data)
-    elif request.method == 'DELETE':
+    
+    def get_serializer_context(self):
+        return {'request': self.request}
+    
+    # def get_queryset(self):
+    #     return Product.objects.select_related('collection').all()[:5]
+
+    # def get_serializer_class(self):
+    #     return ProductSerializer
+
+class ProductDetails(RetrieveUpdateDestroyAPIView):
+    queryset = Product.objects.all()
+    serializer_class = ProductSerializer
+
+    def delete(self, request, pk):
+        product = get_object_or_404(Product, pk=pk)
         if product.orderitems.count() > 0:
             return Response({"error": "Product can not be deleted"}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-    # try:
-    #     product = Product.objects.get(pk=id)
-    #     serializer = ProductSerializer(product)
-    #     print(serializer.data)
-    #     return Response(serializer.data)
-    # except Product.DoesNotExist:
-    #     return Response(status=status.HTTP_404_NOT_FOUND)
 
-@api_view(['GET', 'POST'])
-def collection_list(request):
-    if request.method == 'GET':
-        queryset = Collection.objects.annotate(products_count=Count('products')).all()
-        serializer = CollectionSerializer(queryset, many=True)
-        return Response(serializer.data)
-    elif request.method == 'POST':
-        serializer = CollectionSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data,status=status.HTTP_201_CREATED)
+
+class CollectionList(ListCreateAPIView):
+    queryset = Collection.objects.annotate(products_count=Count('products')).all()
+    serializer_class = CollectionSerializer
+
+
+    def get_serializer_context(self):
+        return {'request': self.request}
+
+
 
 @api_view(['GET', 'PUT', 'DELETE'])
 def collection_detail(request, pk):
@@ -73,5 +59,86 @@ def collection_detail(request, pk):
         serializer.save()
         return Response(serializer.data)
     elif request.method == 'DELETE':
+        if collection.products.count() > 0:
+            return Response({"error": "Collection can not be deleted."}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+
+# class ProductList(APIView):
+
+#     def get(self, request):
+#         queryset = Product.objects.select_related('collection').all()[:5]
+#         serializer = ProductSerializer(queryset, many=True, context={'request': request})
+#         return Response(serializer.data)
+
+#     def post(self, request):
+#         serializer = ProductSerializer(data=request.data)
+#         serializer.is_valid(raise_exception=True)
+#         print(request.data)
+#         print(serializer.validated_data)
+#         serializer.save()
+#         return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+
+# class ProductDetails(APIView):
+
+    # def get(self, request, id):
+    #     product = get_object_or_404(Product, pk=id)
+    #     serializer = ProductSerializer(product)
+    #     print(serializer.data)
+    #     return Response(serializer.data)
+
+    # def put(self, request, id):
+    #     product = get_object_or_404(Product, pk=id)
+    #     serializer = ProductSerializer(product, data=request.data)
+    #     serializer.is_valid(raise_exception=True)
+    #     serializer.save()
+    #     return Response(serializer.data)
+
+    # def delete(self, request, pk):
+    #     product = get_object_or_404(Product, pk=pk)
+    #     if product.orderitems.count() > 0:
+    #         return Response({"error": "Product can not be deleted"}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
+    #     product.delete()
+    #     return Response(status=status.HTTP_204_NO_CONTENT)
+
+# @api_view(['GET', 'PUT', 'DELETE'])
+# def product_details(request, id):
+#     product = get_object_or_404(Product, pk=id)
+#     if request.method == 'GET':
+#         serializer = ProductSerializer(product)
+#         print(serializer.data)
+#         return Response(serializer.data)
+#     elif request.method == 'PUT':
+#         serializer = ProductSerializer(product, data=request.data)
+#         serializer.is_valid(raise_exception=True)
+#         serializer.save()
+#         return Response(serializer.data)
+#     elif request.method == 'DELETE':
+#         if product.orderitems.count() > 0:
+#             return Response({"error": "Product can not be deleted"}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
+#         product.delete()
+#         return Response(status=status.HTTP_204_NO_CONTENT)
+    # try:
+    #     product = Product.objects.get(pk=id)
+    #     serializer = ProductSerializer(product)
+    #     print(serializer.data)
+    #     return Response(serializer.data)
+    # except Product.DoesNotExist:
+    #     return Response(status=status.HTTP_404_NOT_FOUND)
+
+
+
+# @api_view(['GET', 'POST'])
+# def collection_list(request):
+#     if request.method == 'GET':
+#         queryset = Collection.objects.annotate(products_count=Count('products')).all()
+#         serializer = CollectionSerializer(queryset, many=True)
+#         return Response(serializer.data)
+#     elif request.method == 'POST':
+#         serializer = CollectionSerializer(data=request.data)
+#         serializer.is_valid(raise_exception=True)
+#         serializer.save()
+#         return Response(serializer.data,status=status.HTTP_201_CREATED)
